commit 21:00

+++ apps/mail/views/mail-index.jsx
@@ -11,13 +11,15 @@ import { mailService } from "../services/mail.service.js"
 
 export function MailIndex() {
 
+    const [filterBy, setFilterBy] = useState('inbox')
     const [onShow, setOnShow] = useState(false)
     const [mails, setMails] = useState(null)
 
 
     useEffect(() => {
+        console.log('loading')
         loadMails()
-    }, [])
+    }, [filterBy])
 
 
     function onNewMail() {
@@ -25,17 +27,32 @@ export function MailIndex() {
     }
 
     function loadMails() {
-        mailService.query().then(mails => setMails(mails))
+        mailService.query(filterBy).then(mails => setMails(mails))
     }
 
-    function onRemoveMail(mailId) {
-        mailService.remove(mailId).then(() => {
+    function onSetFilter(filterBy) {
+        setFilterBy(filterBy)
+    }
+
+    function onRemoveMail(mail) {
+        const mailId = mail.id
+        if(mail.onTrash) {
+            mailService.remove(mailId).then(() => {
+                const updateMails = mails.filter(mail => mail.id !== mailId)
+                setMails(updateMails)
+            })
+        } else {
+            console.log(mails)
+            mail.onTrash = true
+            mailService.save(mail)
             const updateMails = mails.filter(mail => mail.id !== mailId)
             setMails(updateMails)
-        })
+        }
+
     }
 
     if (!mails) return <h1>Loading...</h1>
+    
     // return <Fragment>
     //     <MailHeader />
     //     <div className='mail-container'>
@@ -47,14 +64,15 @@ export function MailIndex() {
     //         <NewMail onShow={onShow} /> */}
     //     </div>
     // </Fragment>
-    return <Fragment>
+    return <div className='mail-app'>
+
         <MailHeader />
         <div className='mail-container'>
-            <MailNav onNewMail={onNewMail} />
+            <MailNav onNewMail={onNewMail} onSetFilter={onSetFilter} />
             <MailList onShow={!onShow} mails={mails} onRemoveMail={onRemoveMail} />
             <NewMail onShow={onShow} />
         </div>
-    </Fragment>
+    </div>
 
 }
